Fixed launching of locally installed phantomjs
While the installed browser detection already added the node modules of the current
folder to the `PATH`. The launcher itself didn’t and could only start globally
installed instances.
