+++ lib/fileutils.js
@@ -3,19 +3,8 @@ var path = require('path');
 var exec = require('child_process').exec;
 var async = require('async');
 var fileExists = fs.exists || path.exists;
-var _ = require('lodash');
 
-var PATH = 'PATH';
-
-// windows calls it's path 'Path' usually, but this is not guaranteed.
-if (process.platform === 'win32') {
-  PATH = 'Path';
-  Object.keys(process.env).forEach(function(e) {
-    if (e.match(/^PATH$/i)) {
-      PATH = e;
-    }
-  });
-}
+var envWithLocalPath = require('./env-with-local-path');
 
 exports.fileExists = fileExists;
 
@@ -56,13 +45,4 @@ function which(exe, cb) {
   exec('which ' + exe, { env: envWithLocalPath() }, function(err, exePath) {
     cb(!!exePath);
   });
-}
-
-var modulesPath = path.join(process.cwd(), 'node_modules', '.bin');
-
-function envWithLocalPath() {
-  var env = _.clone(process.env);
-  env[PATH] = [modulesPath, env[PATH]].join(process.platform === 'win32' ? ';' : ':');
-
-  return env;
 }
