+++ lib/launcher.js
@@ -5,6 +5,7 @@ var log = require('npmlog');
 var fileutils = require('./fileutils');
 var async = require('async');
 var template = require('./strutils').template;
+var envWithLocalPath = require('./env-with-local-path');
 
 function Launcher(name, settings, config) {
   this.name = name;
@@ -54,10 +55,14 @@ Launcher.prototype = {
   doLaunch: function(cb) {
     var settings = this.settings;
     var self = this;
-    var options = {};
+    var options = {
+      env: envWithLocalPath()
+    };
+
     if (settings.cwd) {
       options.cwd = settings.cwd;
     }
+
     if (settings.exe) {
       var spawn = function spawn(exe, useCrossSpawn) {
         var args = self.getArgs();
